refactor: unify DeviceUnreachableException across Android and iOS
Replace the two separate DeviceUnreachableException types (one in
maestro.android extending IOException, one in maestro extending
RuntimeException for iOS) with a single maestro.DeviceUnreachableException that
both platforms throw.

The unified class extends IOException — required by the Android connection
layer's catch(IOException)=transport-death design, and safe for iOS where no
device-call path catches IOException generically. It carries operation + cause,
plus an optional DeviceDiagnostics (moved into the maestro package) that the
Android path populates for a richer message; iOS leaves it null.

DeviceServerDiedException (Android MODE 1) stays Android-only and now imports
the shared DeviceDiagnostics. No upstream catch sites key off these types —
Orchestra keys retry/skip off MaestroException — so propagation is unchanged.
